backend/requirements.txt
anyio==4.10.0
asgiref==3.9.1
certifi==2025.8.3
Django==4.2.16
django-cors-headers==4.7.0
djangorestframework==3.16.1
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.10
psycopg2-binary==2.9.10
python-dotenv==1.1.1
python-telegram-bot==20.3
sniffio==1.3.1
sqlparse==0.5.3
typing_extensions==4.14.1
